Minor refactor following Package Oriented Design.
