+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Minor refactor following Package Oriented Design. Dependency Injection. [#5](https://github.com/marcosstupnicki/go-users/pull/5)
 - Use bcrypt to hash user password. [#4](https://github.com/marcosstupnicki/go-users/pull/4)
 - Added functionality to load configurations from a struct. [#3](https://github.com/marcosstupnicki/go-users/pull/3)
 - Updated go-webapp version with v1.4.0. [#2](https://github.com/marcosstupnicki/go-users/pull/2)
