--- internal/db/db.go
@@ -1,19 +1,0 @@
-package db
-
-import (
-	"fmt"
-	"github.com/marcosstupnicki/go-users/internal/config"
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
-)
-
-func Connect(cfg config.Database) (*gorm.DB, error){
-	// refer https://github.com/go-sql-driver/mysql#dsn-data-source-name for details
-	dsn := fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?charset=utf8mb4&parseTime=True&loc=Local", cfg.User, cfg.Password, cfg.Host, cfg.Port, cfg.Name)
-	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
-	if err != nil {
-		return nil, err
-	}
-
-	return db, nil
-}
