+++ internal/platform/config/config.go
@@ -3,6 +3,7 @@ package config
 import (
 	"errors"
 	gowebapp "github.com/marcosstupnicki/go-webapp/pkg"
+	"gorm.io/gorm/logger"
 )
 
 var _configs = map[string]Config{
@@ -13,6 +14,7 @@ var _configs = map[string]Config{
 			Host: "127.0.0.1",
 			Port: "3306",
 			Name: "users",
+			LogLevel: logger.Info,
 		},
 	},
 }
@@ -21,7 +23,7 @@ var _configs = map[string]Config{
 func GetConfigFromEnvironment(scope gowebapp.Scope) (Config, error) {
 	config, found := _configs[scope.Environment]
 	if !found  {
-		return Config{},NewErrorConfig(errors.New("config not found"), "config not found for indicated scope")
+		return Config{}, NewErrorConfig(errors.New("config not found"), "config not found for indicated scope")
 	}
 
 	return config, nil
