+++ internal/platform/config/errors.go
