+++ internal/platform/db/db.go
@@ -1,0 +1,31 @@
+package db
+
+import (
+	"fmt"
+	"github.com/marcosstupnicki/go-users/internal/platform/config"
+	"gorm.io/driver/mysql"
+	"gorm.io/gorm"
+	"gorm.io/gorm/logger"
+	"log"
+	"os"
+)
+
+func Connect(cfg config.Database) (*gorm.DB, error) {
+	newLogger := logger.New(
+		log.New(os.Stdout, "\r\n", log.LstdFlags),
+		logger.Config{
+			LogLevel: cfg.LogLevel,
+		},
+	)
+
+	// refer https://github.com/go-sql-driver/mysql#dsn-data-source-name for details
+	dsn := fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?charset=utf8mb4&parseTime=True&loc=Local", cfg.User, cfg.Password, cfg.Host, cfg.Port, cfg.Name)
+	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{
+		Logger: newLogger,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return db, nil
+}
